Refactor - Extraer eventos programados a otra APP
